+++ sdk/resourcemanager/azure-resourcemanager-compute/src/main/java/com/azure/resourcemanager/compute/models/VirtualMachineScaleSet.java
@@ -410,7 +410,7 @@ public interface VirtualMachineScaleSet
     interface DefinitionShared
         extends DefinitionStages.Blank,
             DefinitionStages.WithGroup,
-            DefinitionStages.WithSkuOrFlexibleOrchestrationMode,
+            DefinitionStages.WithSku,
             DefinitionStages.WithProximityPlacementGroup,
             DefinitionStages.WithNetworkSubnet,
             DefinitionStages.WithPrimaryInternetFacingLoadBalancer,
@@ -468,13 +468,7 @@ public interface VirtualMachineScaleSet
         }
 
         /** The stage of a virtual machine scale set definition allowing to specify the resource group. */
-        interface WithGroup extends GroupableResource.DefinitionStages.WithGroup<WithSkuOrFlexibleOrchestrationMode> {
-        }
-
-        /**
-         * The stage of a virtual machine scale set definition allowing to specify SKU or orchestration mode of a virtual machine scale set.
-         */
-        interface WithSkuOrFlexibleOrchestrationMode extends WithSku, WithOrchestrationMode {
+        interface WithGroup extends GroupableResource.DefinitionStages.WithGroup<WithSku> {
         }
 
         /**
@@ -501,7 +495,7 @@ public interface VirtualMachineScaleSet
         }
 
         /** The stage of a virtual machine scale set definition allowing to specify SKU for the virtual machines. */
-        interface WithSku {
+        interface WithSku extends WithOrchestrationMode {
             /**
              * Specifies the SKU for the virtual machines in the scale set.
              *